td-shim: split TDX mailbox related code into dedicated module
Split TDX mailbox related code into dedicated module, to isolate
platform dependent code from platfrom independent code.
