+++ rust-tdshim/src/mp.rs
@@ -2,39 +2,10 @@
 //
 // SPDX-License-Identifier: BSD-2-Clause-Patent
 
-extern crate alloc;
-
-use alloc::vec::Vec;
 use core::mem::size_of;
-use core::{cmp::min, slice};
-use core::{convert::TryInto, ops::RangeInclusive};
-use tdx_tdcall::tdx;
 use zerocopy::{AsBytes, FromBytes};
 
-use crate::{
-    acpi::{self, GenericSdtHeader},
-    memslice::{get_mem_slice_mut, SliceType},
-};
-
-const ACCEPT_CHUNK_SIZE: u64 = 0x2000000;
-const ACCEPT_PAGE_SIZE: u64 = 0x200000;
-const PAGE_SIZE_2M: u64 = 0x200000;
-const PAGE_SIZE_4K: u64 = 0x1000;
-
-// The count of AP to wakeup is limited by the heap size
-// that can be used for stack allocation
-// The maximum size of memory used for AP stacks is 30 KB.
-const MAX_WORKING_AP_COUNT: u32 = 15;
-const AP_TEMP_STACK_SIZE: usize = 0x800;
-
-const MP_WAKEUP_COMMAND_NOOP: u32 = 0;
-const MP_WAKEUP_COMMAND_WAKEUP: u32 = 1;
-const MP_WAKEUP_COMMAND_SLEEP: u32 = 2;
-const MP_WAKEUP_COMMAND_ACCEPT_PAGES: u32 = 3;
-const MP_WAKEUP_COMMAND_AVAILABLE: u32 = 4;
-
-const MAILBOX_APICID_INVALID: u32 = 0xffffffff;
-const MAILBOX_APICID_BROADCAST: u32 = 0xfffffffe;
+use crate::acpi::{self, GenericSdtHeader};
 
 // 255 vCPUs needs 2278 bytes, refer to create_madt().
 const MADT_MAX_SIZE: usize = 0xc00;
@@ -45,239 +16,6 @@ const ACPI_1_0_IO_APIC: u8 = 0x01;
 const ACPI_1_0_INTERRUPT_SOURCE_OVERRIDE: u8 = 0x02;
 const ACPI_1_0_LOCAL_APIC_NMI: u8 = 0x04;
 const ACPI_MADT_MPWK_STRUCT_TYPE: u8 = 0x10;
-
-pub mod mailbox {
-    pub type Field = ::core::ops::Range<usize>;
-
-    pub const COMMAND: Field = 0..4;
-    pub const APIC_ID: Field = 4..8;
-    pub const WAKEUP_VECTOR: Field = 0x08..0x10;
-    pub const CPU_ARRIVAL: Field = 0x900..0x904;
-    pub const CPU_EXITING: Field = 0xa00..0xa04;
-
-    pub const FW_ARGS: usize = 0x800;
-}
-
-pub struct MailBox<'a> {
-    buffer: &'a mut [u8],
-}
-
-impl MailBox<'_> {
-    fn read_volatile<T>(src: *const T) -> T {
-        // Safety: mailbox memory is always valid
-        // Mailbox memory should be read/written using volatile,
-        // it may be changed by AP.
-        unsafe { core::ptr::read_volatile(src) }
-    }
-
-    fn write_volatile<T>(dst: *mut T, src: T) {
-        unsafe { core::ptr::write_volatile(dst, src) }
-    }
-
-    pub fn new(buffer: &mut [u8]) -> MailBox {
-        MailBox { buffer }
-    }
-
-    pub fn apic_id(&self) -> u32 {
-        let p_apic_id = self.buffer[mailbox::APIC_ID].as_ptr() as *const u32;
-        MailBox::read_volatile(p_apic_id)
-    }
-
-    pub fn fw_arg(&self, index: usize) -> u64 {
-        let offset = mailbox::FW_ARGS + index * 8;
-        let p_fw_arg = self.buffer[offset..offset + 8].as_ptr() as *const u64;
-        MailBox::read_volatile(p_fw_arg)
-    }
-
-    pub fn cpu_arrival(&self) -> u32 {
-        let p_cpu_arrival = self.buffer[mailbox::CPU_ARRIVAL].as_ptr() as *const u32;
-        MailBox::read_volatile(p_cpu_arrival)
-    }
-
-    pub fn cpu_exiting(&self) -> u32 {
-        let p_cpu_exiting = self.buffer[mailbox::CPU_EXITING].as_ptr() as *const u32;
-        MailBox::read_volatile(p_cpu_exiting)
-    }
-
-    pub fn set_command(&mut self, command: u32) {
-        let p_command = self.buffer[mailbox::COMMAND].as_ptr() as *mut u32;
-        MailBox::write_volatile(p_command, command);
-    }
-
-    pub fn set_apic_id(&mut self, apic_id: u32) {
-        let p_apic_id = self.buffer[mailbox::APIC_ID].as_ptr() as *mut u32;
-        MailBox::write_volatile(p_apic_id, apic_id);
-    }
-
-    pub fn set_wakeup_vector(&mut self, wakeup_vector: u32) {
-        let p_wakeup_vector = self.buffer[mailbox::WAKEUP_VECTOR].as_ptr() as *mut u32;
-        MailBox::write_volatile(p_wakeup_vector, wakeup_vector);
-    }
-
-    pub fn set_fw_arg(&mut self, index: usize, fw_arg: u64) {
-        let offset = mailbox::FW_ARGS + index * 8;
-        let p_fw_arg = self.buffer[offset..offset + 8].as_ptr() as *mut u64;
-        MailBox::write_volatile(p_fw_arg, fw_arg);
-    }
-}
-
-fn cpu_pause() {
-    unsafe {
-        asm!("pause");
-    }
-}
-
-fn make_apic_range(end: u32) -> RangeInclusive<u32> {
-    // 0 is the bootstrap processor running this code
-    let start = 1;
-
-    RangeInclusive::new(start, end)
-}
-
-// Wait for AP to responde the command set by BSP if needed.
-// Typically AP will set the APIC ID field in mailbox to be invalid
-fn wait_for_ap_responde(mail_box: &mut MailBox) {
-    loop {
-        if mail_box.apic_id() == MAILBOX_APICID_INVALID {
-            mail_box.set_command(MP_WAKEUP_COMMAND_NOOP);
-            break;
-        } else {
-            cpu_pause();
-        }
-    }
-}
-
-// Wait for APs to arrive by checking if they are available
-fn wait_for_ap_arrive(ap_num: u32) {
-    // TODO: Safe because we are the consumer?
-    let mut mail_box = unsafe { MailBox::new(get_mem_slice_mut(SliceType::MailBox)) };
-    for i in make_apic_range(ap_num) {
-        mail_box.set_apic_id(i);
-        mail_box.set_command(MP_WAKEUP_COMMAND_AVAILABLE);
-        wait_for_ap_responde(&mut mail_box);
-    }
-    x86::fence::mfence();
-}
-
-pub fn ap_assign_work(apic_id: u32, stack: u64, entry: u32) {
-    // TODO: safe because we are the consumer?
-    let mut mail_box = unsafe { MailBox::new(get_mem_slice_mut(SliceType::MailBox)) };
-    mail_box.set_wakeup_vector(entry);
-    mail_box.set_fw_arg(0, stack);
-
-    mail_box.set_apic_id(apic_id);
-    mail_box.set_command(MP_WAKEUP_COMMAND_ACCEPT_PAGES);
-
-    wait_for_ap_responde(&mut mail_box);
-}
-
-fn td_accept_pages(address: u64, pages: u64, page_size: u64) {
-    for i in 0..pages {
-        let mut accept_addr = address + i * page_size;
-        let accept_level = if page_size == PAGE_SIZE_2M { 1 } else { 0 };
-        let res = tdx::tdcall_accept_page(accept_addr | accept_level).map_err(|e| {
-            if e == tdx::TdCallError::TdxExitReasonPageSizeMismatch {
-                if page_size == PAGE_SIZE_4K {
-                    log::error!(
-                        "Accept Page Error: 0x{:x}, page_size: {}\n",
-                        accept_addr,
-                        page_size
-                    );
-                } else {
-                    td_accept_pages(accept_addr, 512, PAGE_SIZE_4K);
-                }
-            }
-        });
-    }
-}
-
-fn parallel_accept_memory(apic_id: u64) {
-    // TODO: safe because we are the only user?
-    let mail_box = unsafe { MailBox::new(get_mem_slice_mut(SliceType::MailBox)) };
-
-    // The cpu number, start and end address of memory to be accepted is
-    // set to mailbox fw arguments by mp_accept_memory_resource_range()
-    let cpu_num = mail_box.fw_arg(1);
-    let start = mail_box.fw_arg(2);
-    let end = mail_box.fw_arg(3);
-
-    let stride = ACCEPT_CHUNK_SIZE * cpu_num;
-    let mut phys_addr = start + ACCEPT_CHUNK_SIZE * apic_id;
-
-    while phys_addr < end {
-        let page_num = min(ACCEPT_CHUNK_SIZE, end - phys_addr) / ACCEPT_PAGE_SIZE;
-        td_accept_pages(phys_addr, page_num, ACCEPT_PAGE_SIZE);
-        phys_addr += stride;
-    }
-}
-
-pub fn mp_accept_memory_resource_range(mut cpu_num: u32, address: u64, size: u64) {
-    log::info!(
-        "mp_accept_memory_resource_range: 0x{:x} - 0x{:x} ... (wait for seconds)\n",
-        address,
-        size
-    );
-
-    let active_ap_cnt = if cpu_num - 1 > MAX_WORKING_AP_COUNT {
-        MAX_WORKING_AP_COUNT
-    } else {
-        cpu_num - 1
-    };
-
-    let mut align_low = if address & (ACCEPT_PAGE_SIZE - 1) == 0 {
-        0
-    } else {
-        min(size, ACCEPT_PAGE_SIZE - (address & (ACCEPT_PAGE_SIZE - 1)))
-    };
-    let mut major_part = size - align_low;
-    let mut align_high = 0u64;
-
-    if size > ACCEPT_PAGE_SIZE {
-        major_part = (size - align_low) & !(ACCEPT_PAGE_SIZE - 1);
-        if major_part < ACCEPT_PAGE_SIZE {
-            align_low += major_part;
-            major_part = 0;
-        } else {
-            align_high = size - align_low - major_part;
-        }
-    }
-
-    wait_for_ap_arrive(active_ap_cnt);
-
-    let mut stacks: Vec<u8> = Vec::with_capacity(AP_TEMP_STACK_SIZE * active_ap_cnt as usize);
-    // TODO: safe because we are the only user?
-    let mut mail_box = unsafe { MailBox::new(get_mem_slice_mut(SliceType::MailBox)) };
-
-    // BSP calles the same function parallel_accept_memory to accept memory,
-    // so set the firmware arguments here.
-    // To do: Set these parameter only in ap_assign_work() when there's
-    // multiple cpus.
-    mail_box.set_fw_arg(1, active_ap_cnt as u64 + 1);
-    mail_box.set_fw_arg(2, address + align_low);
-    mail_box.set_fw_arg(3, address + size);
-
-    if major_part > 0 {
-        // 0 is the bootstrap processor running this code
-        for i in make_apic_range(active_ap_cnt) {
-            let ap_stack = stacks.as_ptr() as u64 + (i - 1) as u64 * 0x800;
-            ap_assign_work(i, ap_stack, parallel_accept_memory as *const () as u32);
-        }
-    }
-
-    parallel_accept_memory(0);
-
-    td_accept_pages(address, align_low / PAGE_SIZE_4K, PAGE_SIZE_4K);
-
-    td_accept_pages(
-        address + align_low + major_part,
-        align_high / PAGE_SIZE_4K,
-        PAGE_SIZE_4K,
-    );
-
-    wait_for_ap_arrive(active_ap_cnt);
-
-    log::info!("mp_accept_memory_resource_range: done\n");
-}
 
 pub struct Madt {
     pub data: [u8; MADT_MAX_SIZE],
