+++ rust-tdshim/src/td/mod.rs
@@ -3,5 +3,6 @@
 // SPDX-License-Identifier: BSD-2-Clause-Patent
 
 mod tdx;
+mod tdx_mailbox;
 
 pub use tdx::*;
