+++ rust-tdshim/src/td/tdx.rs
@@ -44,3 +44,7 @@ pub fn enable_execution_disable_bit() {
 pub fn get_shared_page_mask() -> u64 {
     tdx_tdcall::tdx::td_shared_page_mask()
 }
+
+pub fn accept_memory_resource_range(mut cpu_num: u32, address: u64, size: u64) {
+    super::tdx_mailbox::accept_memory_resource_range(cpu_num, address, size)
+}
